requirement.txt
pycryptodome
asyncio_dgram
asyncio
aiohttp
scapy